refactor: moving the mysql client
