+++ azurerm/internal/provider/services.go
@@ -38,6 +38,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/monitor"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/msi"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/mssql"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/mysql"
 )
 
 func SupportedServices() []common.ServiceRegistration {
@@ -78,5 +79,6 @@ func SupportedServices() []common.ServiceRegistration {
 		monitor.Registration{},
 		msi.Registration{},
 		mssql.Registration{},
+		mysql.Registration{},
 	}
 }
