+++ azurerm/internal/services/mysql/client/client.go
@@ -1,4 +1,4 @@
-package mysql
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/mysql/mgmt/2017-12-01/mysql"
@@ -13,7 +13,7 @@ type Client struct {
 	VirtualNetworkRulesClient *mysql.VirtualNetworkRulesClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	ConfigurationsClient := mysql.NewConfigurationsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&ConfigurationsClient.Client, o.ResourceManagerAuthorizer)
